changes see summary:
code cleanup,
