+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.Api/HelperFunctions/JwtMiddleware.cs
@@ -1,28 +1,28 @@
-﻿using AutoBuildApp.Security.Enumerations;
-using AutoBuildApp.Security.FactoryModels;
-using AutoBuildApp.Security.Interfaces;
-using AutoBuildApp.Security.Models;
+﻿using AutoBuildApp.Security;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Net.Http.Headers;
 using System;
 using System.Linq;
 using System.Net.Http.Headers;
 using System.Security.Claims;
-using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
 
 
 namespace AutoBuildApp.Api.HelperFunctions
 {
+    /// <summary>
+    /// Custom middleware into the request pipeline of ASP.NET Core application.
+    /// </summary>
     public class JwtMiddleware
     {
         /// <summary>
-        /// processes http requests.
+        ///  in order to execute next middleware in a sequence, 
+        ///  it should have RequestDelegate type parameter in the constructor. 
         /// </summary>
         private readonly RequestDelegate _next;
-        private JWTValidator _validateAuthorizationHeader;
+
+        private JwtValidator _validateAuthorizationHeader;
 
         // i want to store this into the 
         // http.context.item["ClaimsPrincipal"] =.
@@ -36,16 +36,19 @@ namespace AutoBuildApp.Api.HelperFunctions
         {
             _next = next;
         }
+
+        /// <summary>
+        ///   not asynchronous will block the 
+        ///  thread till the time it completes the execution. 
+        ///  So, make it asynchronous by using async and await
+        ///  to improve performance and scalability. 
+        /// </summary>
+        /// <param name="httpContext"></param>
+        /// <returns></returns>
         public async Task InvokeAsync(HttpContext httpContext)
         {
-            /// postman isnt sending a request header containing aa referer....
-            /// so:
-            /// https://learning.postman.com/docs/sending-requests/requests/
-            /// 
-            ///
 
-            Console.WriteLine($"\n\t " +
-                $"IN THE JWT MIDDLE WARE \n");
+            Console.WriteLine($"\n\t " + $"IN THE JWT MIDDLE WARE \n");
 
             //STEP 1: EXTRACT THE TOKEN (IF EXISTS ANY)
             #region EXTRACT THE TOKEN (IF EXISTS ANY)
@@ -55,8 +58,10 @@ namespace AutoBuildApp.Api.HelperFunctions
                 .Split(" ")
                 .Last();
 
-            if( token!= null) { Console.WriteLine($"Token to be validated : {token } \n"); }
-            Console.WriteLine($"Token to be validated : {token } \n");
+            if (token != null)
+            {
+                Console.WriteLine($"Token to be validated : {token } \n");
+            }
 
             var authHeader = httpContext.Request.Headers["Authorization"];//[0];
 
@@ -66,16 +71,14 @@ namespace AutoBuildApp.Api.HelperFunctions
                 var scheme = headerValue.Scheme;
 
                 var parameter = headerValue.Parameter;
-                Console.WriteLine($"\n\t " +
-                    $"IN THE TRY PARSE \n" +
-                    $" scheme : { scheme}\n" +
-                    $" parameter :{ parameter }\n");
+                Console.WriteLine($"\n\t " + $"IN THE TRY PARSE \n" + $" scheme : { scheme}\n" + $" parameter :{ parameter }\n");
 
                 // scheme will be "Bearer"
                 // parmameter will be the token itself.
                 token = parameter;
             }
             #endregion
+
             //STEP 2: EXTRACT THE REFERER
             #region EXTRACT THE REFERER
             var Url =
@@ -108,35 +111,40 @@ namespace AutoBuildApp.Api.HelperFunctions
                 Console.WriteLine($"YOU ARE REQUESTING THE AUTHDEMO URL\n");
             }
 
+            /*
+             * 2) user is NOT logged in and are not authorized -> 
+             * *status code: 401.here comes in the REDIRECT URL.
+             * They are directed to the login page and later
+             * to the page intended(if they are aurthorized)
+             *
+             */
 
-            if (token != null && token.Length != 0) // there is a JWT token 
+            if (token != null && token.Length != 0) // THERE EXISTS A JWT token 
             {
-                _validateAuthorizationHeader = new JWTValidator(token); // validate the token 
+                _validateAuthorizationHeader = new JwtValidator(token); // validate the token 
                 var result = ValidateTheToken(httpContext, token);
                 if (result == false) //IF JWT NOT VALID
                 {
-                    Console.WriteLine($"THE TOKEN PASSED IS NOT VALID JWT!");
+                    //Console.WriteLine($"THE TOKEN PASSED IS NOT VALID JWT!");
+                    httpContext.Response.StatusCode = StatusCodes.Status403Forbidden;
                     return;
-
-                   // return Task.FromResult( await httpContext.Response.WriteAsync("Hello World! "+ httpContext.Response.StatusCode));
-                   // await httpContext .Response .WriteAsync(httpContext.Response.StatusCode.ToString());
 
                 }
             }
             else if (token == null || token == " ") // else set the default principle:
             {
-                if (Thread.CurrentPrincipal != null) // this check may be removed...
-                {
-                    Thread.CurrentPrincipal = (ClaimsPrincipal)Thread.CurrentPrincipal;
-                }
-                else { DefaultClaimsPrinciple(); }
+
+                DefaultClaimsPrinciple();
+
+                //if (Thread.CurrentPrincipal != null) // this check may be removed...
+                //{
+                //    Thread.CurrentPrincipal = (ClaimsPrincipal)Thread.CurrentPrincipal;
+                //}
+
+               
             }
 
-
-
-            Console.WriteLine($"\n\t " +
-                $"END OF THE JWT MIDDLE WARE \n");
-
+            Console.WriteLine($"\n\t " +$"END OF THE JWT MIDDLE WARE \n");
             await _next(httpContext);
         }
 
@@ -145,32 +153,23 @@ namespace AutoBuildApp.Api.HelperFunctions
 
             Console.WriteLine($"VALIDATING THE TOKEN METHOD.");
             ///https://dev.to/tjindapitak/better-way-of-storing-per-request-data-across-middlewares-in-asp-net-core-1m9k
-            ///
             if (!_validateAuthorizationHeader.IsValidJWT()) // JWT IS NOT VALID, END CALL
             {
 
-                httpContext.Response.StatusCode = StatusCodes.Status404NotFound;//400; //Bad Request   
-                
+                httpContext.Response.StatusCode = StatusCodes.Status404NotFound;//400; //Bad Request 
                 return false;
             }
             else //THE JWT IS VALID, THEREFORE SET THE CLAIMS PRINCIPLE TO THREAD.
             {
                 #region IF THE AUTH HEADER CONTAINS VALID TOKEN  -> SET CLAIMSPRINCIPAL TO THREAD
+                JwtParser jwtParser = new JwtParser(token);
 
-                var userPrinciple =
-                    _validateAuthorizationHeader.ParseForClaimsPrinciple();
-
+                var userPrinciple = jwtParser.ParseForClaimsPrinciple();
                 _threadPrinciple = (ClaimsPrincipal)Thread.CurrentPrincipal; // get the principle on the thread
-                Console.WriteLine($"\nIN THE JWT MIDDLEWARE CHEWCKING THE PRINCIPLE NAME: {_threadPrinciple.Identity.Name}\n");
-
-
-
+                // Console.WriteLine($"\nIN THE JWT MIDDLEWARE CHEWCKING THE PRINCIPLE NAME: {_threadPrinciple.Identity.Name}\n");
                 Thread.CurrentPrincipal = _threadPrinciple; // SETTING THE PARSED TOKEN, TO THE THREAD.
 
-               
                 #endregion
-
-                // here i will be 
 
 
             }
