+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.DataAccess/DAOGlobals/DAOGlobal.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace AutoBuildApp.DataAccess.DAOGlobals
+{
+    public static class DAOGlobal
+    {
+
+
+        public const int TIMEOUT_SHORT = 5;
+        public const int TIMEOUT_MEDIUM = 30;
+        public const int TIMEOUT_LONG = 60;
+
+    }
+}
