+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.DataAccess/SessionsDAO.cs
@@ -59,7 +59,7 @@ namespace AutoBuildApp.DataAccess
                         #region SQL related
 
                         // https://learning.oreilly.com/library/view/adonet-in-a/0596003617/ch04s05.html
-                        command.CommandTimeout = TimeSpan.FromSeconds(DAOGlobals.TIMEOUT_LONG).Seconds;
+                        command.CommandTimeout = TimeSpan.FromSeconds(DAOGlobals2.TIMEOUT_LONG).Seconds;
                         // 1) Create a Command, and set its CommandType property to StoredProcedure.
                         command.CommandType = CommandType.StoredProcedure;
                         // 2) Set the CommandText to the name of the stored procedure.
