+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.Managers/FeatureManagers/AnalyticsManager.cs
@@ -123,7 +123,7 @@ namespace AutoBuildApp.Managers.FeatureManagers
         //    return analyticCharts;
         //}
 
-
+        // TURN INTO ASYNC 
         public void GetChartData(int graphType)
         {
             DBViews specifiedChart = (DBViews)graphType;
