+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.Security/Models/Claims.cs
@@ -31,8 +31,6 @@ namespace AutoBuildApp.Security.Models
             {
                 return false;
             }
-
-
             return 
                 this.Permission == other.Permission 
                 && this.scopeOfPermissions == other.scopeOfPermissions;
@@ -40,96 +38,5 @@ namespace AutoBuildApp.Security.Models
 
     }
 
-    /// <summary>
-    /// the claims per role class set 
-    /// predefined claims per Roles: admin, basic, unregistered, vendor and developer
-    /// to allow consistency for the set of claims defined per user
-    /// </summary>
-    public sealed class ClaimsPerRoles : IEquatable<ClaimsPerRoles>
-    {
-        public static readonly IList<Claims>
-            claimsPerAdmin = new List<Claims>()
-            {
-            new Claims("all","all"),
-            };
-        public static readonly IList<Claims> 
-            claimsPerVendor = new List<Claims>()
-            {
-            new Claims("Read Only","AutoBuild"),
-                 new Claims("Delete","self"),
-                 new Claims("Update","self"),
-                 new Claims("Edit","self"),
-                 new Claims("Create" ,"reviews"),
-                 new Claims("Delete" ,"selfReview"),
-                 new Claims("Update" ,"selfReview"),
-                 new Claims("Create","Products"),
-                new Claims("Update","VendorProducts"),
-                new Claims("Delete","VendorProductsOnly"),
-                new Claims("Update","VendorProducts")
-
-            };
-        public static readonly IList<Claims> 
-            claimsPerBasic = new List<Claims>()
-            {
-                new Claims("Read Only","AutoBuild"),
-                 new Claims("Delete","self"),
-                 new Claims("Update","self"),
-                 new Claims("Edit","self"),
-                 new Claims("Create" ,"reviews"),
-                 new Claims("Delete" ,"selfReview"),
-                 new Claims("Update" ,"selfReview")
-            };
-        public static readonly IList<Claims> 
-            claimsPerUnregistered = new List<Claims>()
-            {
-                new Claims("Read Only","AutoBuild")
-            };
-        public static readonly IList<Claims> 
-            claimsPerDevolper = new List<Claims>()
-            {
-                new Claims("Read Only","AutoBuild")
-            };
-
-        /// <summary>
-        /// here I define static claims per roles 
-        /// this makes it easier
-        /// </summary>
-        public static readonly ClaimsPerRoles AdminUserClaims
-            = new ClaimsPerRoles(claimsPerAdmin);
-        public static readonly ClaimsPerRoles BasicUserClaims
-            = new ClaimsPerRoles(claimsPerBasic);
-        public static readonly ClaimsPerRoles VendorUserClaims
-            = new ClaimsPerRoles(claimsPerVendor);
-        public static readonly ClaimsPerRoles UnregisteredUserClaims
-            = new ClaimsPerRoles(claimsPerUnregistered);
-        public IEnumerable<Claims> Permissions { get; set; }
-
-        private ClaimsPerRoles(IEnumerable<Claims> PermissionsPassed)
-        {
-            this.Permissions = PermissionsPassed;
-        }
-
-        /// <summary>
-        /// the method is inherited from the IEquatable interface
-        /// whcih returns true if two ClaimsPerRoles objects are equal and 
-        /// false otherwise.
-        /// </summary>
-        /// <param name="other"></param>
-        /// <returns></returns>
-        public bool Equals(ClaimsPerRoles other)
-        {
-            if (other is null)
-            {
-                return false;
-            }
-            if(claimsPerAdmin.Equals(other.Permissions) || claimsPerBasic.Equals(other.Permissions) ||
-                claimsPerDevolper.Equals(other.Permissions) || claimsPerUnregistered.Equals(other.Permissions) ||
-                claimsPerVendor.Equals(other.Permissions)){
-
-                return true;
-            }
-            return false;
-        }
-    }
 
 }
